requirements.txt
streamlit
treamlit-drawable-canvas
Pillow